Particle changes and rotation lerping

+++ source/objects.h
@@ -100,10 +100,10 @@ typedef struct {
 } Color;
 
 typedef struct {
-    u8 r;
-    u8 g;
-    u8 b;
-    u8 a;
+    float r;
+    float g;
+    float b;
+    float a;
 } ColorAlpha;
 
 struct ColorChannel {
